+++ Ostrich/OstrichRenderer/RenderMath/Random.cs
@@ -4,7 +4,7 @@
     {
         public static long Seed = 1;
 
-        public static double Get()
+        public static float Get()
         {
             Seed = (0x5DEECE66DL * Seed + 0xB16) & 0xFFFFFFFFFFFFL;
             return (Seed >> 16) / (float)0x100000000L;
